# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Badminton Court Booking API",
        "description": "API untuk sistem pemesanan lapangan badminton",
        "contact": {
            "name": "Badminton Booking Support",
            "email": "admin@badminton-booking.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/courts": {
            "get": {
                "tags": [
                    "Courts"
                ],
                "summary": "Dapatkan semua lapangan aktif",
                "description": "Mengambil daftar semua lapangan badminton yang memiliki status aktif",
                "operationId": "78c3229f958262505d946a5a3fefa6b3",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data lapangan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Courts retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Court"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Courts"
                ],
                "summary": "Buat lapangan baru",
                "description": "Membuat lapangan badminton baru",
                "operationId": "1a8a5f65baa118090bc3fffdf346d05c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price_per_hour"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nama lapangan",
                                        "type": "string",
                                        "example": "Lapangan A"
                                    },
                                    "description": {
                                        "description": "Deskripsi lapangan",
                                        "type": "string",
                                        "example": "Lapangan premium dengan lantai kayu"
                                    },
                                    "price_per_hour": {
                                        "description": "Harga per jam",
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 50000
                                    },
                                    "status": {
                                        "description": "Status lapangan",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lapangan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Court created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Court"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/courts/{id}": {
            "get": {
                "tags": [
                    "Courts"
                ],
                "summary": "Dapatkan detail lapangan",
                "description": "Mengambil detail lapangan badminton berdasarkan ID",
                "operationId": "82c07587d3f88ff03db56cab8847a6e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID lapangan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail lapangan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Court retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Court"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lapangan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Courts"
                ],
                "summary": "Update lapangan",
                "description": "Mengupdate data lapangan badminton",
                "operationId": "469d33e2bb18d931b55a56805a5a8dad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID lapangan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nama lapangan",
                                        "type": "string",
                                        "example": "Lapangan A Updated"
                                    },
                                    "description": {
                                        "description": "Deskripsi lapangan",
                                        "type": "string",
                                        "example": "Deskripsi lapangan yang diupdate"
                                    },
                                    "price_per_hour": {
                                        "description": "Harga per jam",
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 60000
                                    },
                                    "status": {
                                        "description": "Status lapangan",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lapangan berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Court updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Court"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lapangan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Courts"
                ],
                "summary": "Hapus lapangan",
                "description": "Menghapus lapangan badminton",
                "operationId": "05ffe255617a7b131b14b14cdbaa027a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID lapangan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lapangan berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Court deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lapangan tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Court": {
                "title": "Court",
                "description": "Model lapangan badminton",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Lapangan A"
                    },
                    "description": {
                        "type": "string",
                        "example": "Lapangan premium dengan lantai kayu"
                    },
                    "price_per_hour": {
                        "type": "number",
                        "format": "decimal",
                        "example": 50000
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Error response format",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Error occurred"
                    },
                    "error": {
                        "type": "string",
                        "example": "Detailed error message"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "description": "Validation error response format",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The field is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}